refactor: Util 클래스 대신 인터페이스로 랜덤숫자 생성
