+++ docs/README.md
@@ -37,6 +37,7 @@
 * [ ] 인터페이스 적극 활용(메시지를 던지자)
 * [x] Integer 대신 Long 사용 고려
   * 다시 Integer로 변경
+* [x] Util 클래스 대신 인터페이스로 랜덤숫자 생성
 
 ## 고민해볼 주제
