+++ src/main/java/racingcar/service/NumberGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.service;
+
+public interface NumberGenerator {
+    int generate();
+}
