+++ src/main/java/racingcar/service/RandomNumberGenerator.java
@@ -1,13 +1,12 @@
-package racingcar.util;
+package racingcar.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import racingcar.constant.Rule;
 
-public class RandomNumberGenerator {
-    private RandomNumberGenerator() {
-    }
+public class RandomNumberGenerator implements NumberGenerator {
 
-    public static Integer generate() {
+    @Override
+    public int generate() {
         return Randoms.pickNumberInRange(Rule.MIN_POSSIBILITY, Rule.MAX_POSSIBILITY);
     }
 }
